requirements.txt
python-dotenv
pydantic
pydantic_settings
# For Web
wikipedia
playwright
ddgs 

# For API
fastapi

# For MCP
fastmcp
